requirements.txt
django
python-dotenv
djangorestframework
pillow

# flake8
django-mptt
# django-treebeard

### SWAGGER UI
drf-spectacular

### TESTING
coverage
pytest
pytest-django